requirements.txt
# 基于扩散模型和强化学习的ADC生成挑战赛 - 依赖包

# 核心深度学习框架
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
pytorch-lightning>=1.7.0

# 数据处理和科学计算
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0

# 分子和化学信息学
rdkit>=2022.9.1
mordred>=1.2.0
deepchem>=2.6.0

# 生物信息学
biopython>=1.79
prody>=2.4.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
py3Dmol>=1.8.0

# 机器学习和强化学习
transformers>=4.21.0
stable-baselines3>=1.6.0
gym>=0.21.0

# 工具库
tqdm>=4.64.0
PyYAML>=6.0
jupyter>=1.0.0
ipywidgets>=7.7.0

# 数据库和存储
sqlite3
pickle5
h5py>=3.7.0

# 其他工具
requests>=2.28.0
openpyxl>=3.0.10
psutil>=5.9.0